Add parameter for message size
Added a variable DATA_BYTES to the Makefile which
sets the size of the message that will be transmitted.
This was required in order to compare with ping which has a
default data transmission size of 56 bytes.

Also, I added logic that will throw out results if the client
does not receive all of the data that it sent (i.e. some packets were lost).